moved some files into data folder
