--- app/sponsors/SponsorData.ts
@@ -1,171 +1,0 @@
-import { Sponsor } from "@/types/types";
-
-export const SponsorData: Array<Sponsor> = [
-    {
-        name: "AEZION",
-        logo: "/assets/sponsors/aeizon.webp",
-        link: "https://www.aezion.com/",
-    },
-
-    {
-        name: "Airtech",
-        logo: "/assets/sponsors/airtech.webp",
-        link: "https://airtech.com/",
-    },
-
-    {
-        name: "Al Lamb's Dallas Honda",
-        logo: "/assets/sponsors/allamb.webp",
-        link: "https://dallashonda.com/",
-    },
-    
-    {
-        name: "Altium Design",
-        logo: "/assets/sponsors/altium.webp",
-        link: "https://www.altium.com/",
-    },
-    
-    {
-        name: "ANSYS",
-        logo: "/assets/sponsors/ansys.svg",
-        link: "https://www.ansys.com/",
-    },
-    
-    {
-        name: "CONVERGE CFD",
-        logo: "/assets/sponsors/converge.svg",
-        link: "https://convergecfd.com/",
-    },
-    
-    {
-        name: "Evan Composites",
-        logo: "/assets/sponsors/evancomposites.webp",
-        link: "https://www.evancomposites.com/",
-    },
-    
-    {
-        name: "Fibre Glast",
-        logo: "/assets/sponsors/fibreglast.webp",
-        link: "https://www.fibreglast.com/",
-    },
-    
-    {
-        name: "HEXAGON",
-        logo: "/assets/sponsors/hexagon.svg",
-        link: "https://hexagon.com/",
-    },
-
-    {
-        name: "LabJack",
-        logo: "/assets/sponsors/labjack.webp",
-        link: "https://labjack.com/",
-    },
-    
-    {
-        name: "Liqui Moly",
-        logo: "/assets/sponsors/liquimoly.svg",
-        link: "https://www.liqui-moly.com",
-    },
-    
-    {
-        name: "Micro Measurements",
-        logo: "/assets/sponsors/micromeasurements.svg",
-        link: "https://micro-measurements.com/",
-    },
-    
-    {
-        name: "NTBAA",
-        logo: "/assets/sponsors/ntbaa.svg",
-        link: "https://ntbaaonline.com/",
-    },
-    
-    {
-        name: "Peterbilt",
-        logo: "/assets/sponsors/peterbilt.webp",
-        link: "https://www.peterbilt.com/",
-    },
-    
-    {
-        name: "RedBull",
-        logo: "/assets/sponsors/redbull.webp",
-        link: "https://www.redbull.com/us-en/",
-    },
-
-    {
-        name: "Rescale",
-        logo: "/assets/sponsors/rescale.svg",
-        link: "https://www.rescale.com/",
-    },
-    
-    {
-        name: "Rockwall MakerSpace",
-        logo: "/assets/sponsors/rockwallmakerspace.webp",
-        link: "https://rockwallmakerspace.org/",
-    },
-    
-    {
-        name: "SAE",
-        logo: "/assets/sponsors/sae.svg",
-        link: "https://www.sae.org/",
-    },
-    
-    {
-        name: "Schroth Racing",
-        logo: "/assets/sponsors/schroth.svg",
-        link: "https://www.schrothracing.com/",
-    },
-    
-    {
-        name: "Sherwin Williams",
-        logo: "/assets/sponsors/sherwinwilliams.webp",
-        link: "https://www.sherwin-williams.com/",
-    },
-    
-    {
-        name: "Siemens",
-        logo: "/assets/sponsors/siemens.png",
-        link: "https://www.siemens.com/",
-    },
-    
-    {
-        name: "SKF",
-        logo: "/assets/sponsors/skf_logo.webp",
-        link: "https://www.skf.com/us",
-    },
-    
-    {
-        name: "SolidWorks",
-        logo: "/assets/sponsors/solidworks.svg",
-        link: "https://www.solidworks.com/",
-    },
-    
-    {
-        name: "Tesla",
-        logo: "/assets/sponsors/tesla.webp",
-        link: "https://www.tesla.com/",
-    },
-    
-    {
-        name: "TTA",
-        logo: "/assets/sponsors/tta.svg",
-        link: "https://www.testtechnology.com/",
-    },
-    
-    {
-        name: "UTD",
-        logo: "/assets/sponsors/utd.svg",
-        link: "https://www.utdallas.edu/",
-    },
-    {
-        name: "WelshAdvisors",
-        logo: "/assets/sponsors/welshadvisors.webp",
-        link: "https://www.welshadvisors.com/",
-    },
-    {
-        name: "Omni Powertrain",
-        logo: "/assets/sponsors/omnipowertrain.png",
-        link: "https://www.omnipowertrain.com/",
-    },
-];
-
-export default SponsorData;
